requirements.txt
streamlit
pandas
numpy
scipy.stats
plotly.graph_objects